pyproject.toml
[project]
name = "otus-project-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.1.2",
    "flask==3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.0",
]
